+++ src/main.lua
@@ -443,6 +443,7 @@ local function on_ready_late()
 	import "Scripts/HarvestLogic_Late.lua"
 	import "Scripts/HarvestPresentation_Late.lua"
 	import "Scripts/HubPresentation_Late.lua"
+	import "Scripts/InteractLogic_Late.lua"
 	import "Scripts/MarketLogic_Late.lua"
 	import "Scripts/RoomPresentation_Late.lua"
 	import "Scripts/ShrinePresentation_Late.lua"